fix(fuse): widen the xattr size bound so 32-bit targets compile
`n > math.MaxUint32` does not compile where int is 32 bits wide: the untyped constant does not
fit the type it is compared against. linux/386 and linux/arm failed the cross-build matrix
while every test passed on this 64-bit host, because the defect is a compile error and not a
wrong answer. The comparison now widens through uint64, which is correct on both — where int is
32 bits the branch is unreachable, and where it is 64 bits it is the real bound.

The test had the same bug in the same shape, and the fix there is not just a wider comparison:
`int(uint64(math.MaxUint32))` is still a constant expression, so it fails to compile on a
32-bit target even inside a branch that can never be taken there — constant conversions are
checked before any branch is considered. The bound goes through a uint64 *variable*, and the
two over-MaxUint32 rows are appended conditionally rather than filtered afterwards, because on
a 32-bit platform the arithmetic that produces them collapses onto values the other rows
already use and a value-based filter would let a row through with the wrong expectation.

CI's cross-build comment describes this defect in advance, from #198: `safeIntToUint32` guarded
its conversion with `i > 0xFFFFFFFF` and internal/fuse did not compile for any 32-bit target,
which cost linux/armv7 a place in the release matrix. The 32-bit cells were added then, and the
`go vet` step alongside them exists because `go build` does not compile _test.go files — the
comment names `int(math.MaxUint32)` in a test as the same error in the same shape. Both cells
and both steps did their job here.

Verified: go vet across linux/386, linux/arm, linux/amd64, linux/arm64 and darwin — which
compiles the tests, unlike go build — golangci-lint 0 issues, and the clamp's mutation check
re-run after the rewrite, still failing with xattrSize(4294967296) = 0.

+++ internal/fuse/xattr.go
@@ -130,8 +130,14 @@ func refuseXattrNamespace(name string) bool {
 // value is a caller that sizes a short buffer, gets no ERANGE because the filesystem also compared the
 // wrapped number, and reads a truncated attribute as if it were whole — silent data loss on a read path.
 // MaxUint32 cannot be mistaken for a real attribute size and fails the caller's own allocation instead.
+//
+// The comparison widens to uint64 rather than testing `n > math.MaxUint32` directly. On a 32-bit build an
+// int *is* 32 bits, so that constant does not fit one and the package fails to compile — which is how the
+// first version of this function broke the linux/386 and linux/arm cross-builds while passing every test
+// on a 64-bit host. Widening is correct on both: where int is 32 bits the branch is simply never taken,
+// and where it is 64 bits it is the real bound.
 func xattrSize(n int) uint32 {
-	if n < 0 || n > math.MaxUint32 {
+	if n < 0 || uint64(n) > math.MaxUint32 {
 		return math.MaxUint32
 	}
